V1.5 add filter and improved display

+++ STM32H750_thermal camera/MLX90640/functions/MLX90640_API.c
@@ -265,7 +265,7 @@ int MLX90640_GetCurMode(uint8_t slaveAddr)
 
 //------------------------------------------------------------------------------
 __attribute__((section("RAMCODE")))
-void MLX90640_CalculateTo(uint16_t *frameData, const paramsMLX90640 *params, float emissivity, float tr, float *result)
+void MLX90640_CalculateTo(uint16_t *frameData, const paramsMLX90640 *params, float emissivity, float tr, uint16_t *result)
 {
     float vdd;
     float ta;
