+++ STM32H750_thermal camera/MLX90640/headers/MYMLX90640_API.h
@@ -116,7 +116,7 @@
     float MLX90640_GetTa(uint16_t *frameData, const paramsMLX90640 *params);
     void MLX90640_GetImage(uint16_t *frameData, const paramsMLX90640 *params, float *result);
 		
-    void MLX90640_CalculateTo(uint16_t *frameData, const paramsMLX90640 *params, float emissivity, float tr, float *result);
+    void MLX90640_CalculateTo(uint16_t *frameData, const paramsMLX90640 *params, float emissivity, float tr, uint16_t *result);
 		
     int MLX90640_SetResolution(uint8_t slaveAddr, uint8_t resolution);
     int MLX90640_GetCurResolution(uint8_t slaveAddr);
